add new buy back modal
